Fixed table of contents anchor links
I went to look at the installation instructions this morning and was getting a 404 on the anchor links. If it worked before, Github may have changed something.  Please make sure this works for you in your browser before merging, but this change works for me.  If there's a more succinct way of making these links, I am not familiar.  I'd googled and tried a few things before I figured out what worked.